+++ lib/torch/nn/identity.rb
@@ -1,0 +1,13 @@
+module Torch
+  module NN
+    class Identity < Module
+      def initialize(*args, **options)
+        super()
+      end
+
+      def forward(input)
+        input
+      end
+    end
+  end
+end